Support local deployment and hello-world inner docker.
The demo can be deployed on localhost.
All the ROSE commands are overwritten by running hello-world inner docker.

The inner docker is built inside the outer docker container and runs in rootless mode. The root user in the inner docker is mapped to a non-root user in the outer docker.